Split code into separate files, better reset function
